+++ PSFramework/internal/scriptblocks/strings.validate.ps1
@@ -1,3 +1,3 @@
 ﻿Set-PSFScriptblock -Name 'PSFramework.Validate.SafeName' -Scriptblock {
-	$_ -match '^[\d\w_-\.]+$'
+	$_ -match '^[\d\w_\-\.]+$'
 } -Global